Added compressor (teleop only?) and cargo pneumatics (maybe works?)
Not sure if the compressor block in robot.java OperatorControl() is the best place, we might want to move it somewhere else.

Not sure if the pneumatics actually work. I changed them to DoubleSolenoids (because that's what they are) but it still needs to be tested.

Changed a few constants, and added some TODOs